requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gspread
python-dotenv==1.0.1
pandas==2.2.2